+++ app/package.json
@@ -13,7 +13,7 @@
     "dependencies": {
         "@mobilabs/easing": "^1.0.2",
         "@mysten/bcs": "^1.3.0",
-        "@mysten/dapp-kit": "0.14.21",
+        "@mysten/dapp-kit": "^0.14.49",
         "@mysten/enoki": "^0.4.24",
         "@mysten/sui": "^1.21.1",
         "@radix-ui/colors": "^3.0.0",
