+++ app/src/apps/commander-v2/Playground.tsx
@@ -23,9 +23,21 @@ import { useSuiClient } from "@mysten/dapp-kit";
 import { EventBus, GameEvent } from "./engine/EventBus";
 import { bcs } from "./types/bcs";
 import * as THREE from "three";
+import { NavLink } from "react-router-dom";
 
 export const SIZE = 10;
 export const LS_KEY = "commander-v2";
+
+export type AttackEvent = {
+    game: string;
+    attacker: [number, number];
+    target: [number, number];
+    damage: number;
+    is_dodged: boolean;
+    is_missed: boolean;
+    is_crit: boolean;
+    is_kia: boolean;
+};
 
 /**
  * The main component of the game.
@@ -46,6 +58,7 @@ export function Playground() {
     const localKey = localStorage.getItem(LS_KEY);
     const { data: map, isFetching, isFetched, refetch } = useGame({ id: localKey, enabled: true });
     const [lockedTx, setLockedTx] = useState<Transaction | null>(null);
+    const canTransact = !!zkLogin.address && !!executeTransaction && !isExecuting;
 
     useEffect(() => {
         loadModels().then(() => setModelsLoaded(true));
@@ -62,7 +75,11 @@ export function Playground() {
                 const result = await moveUnit(event.path.map((p: THREE.Vector2) => [p.x, p.y]));
                 if (!result) return console.log("unable to move unit");
                 setLockedTx(result.tx);
-                eventBus.dispatchEvent({ type: "sui", action: "trace:success" });
+                eventBus.dispatchEvent({
+                    type: "sui",
+                    action: "trace",
+                    message: "path checked",
+                });
             }
 
             if (event.action === "aim") {
@@ -72,7 +89,7 @@ export function Playground() {
                 const result = await performAttack([x0, y0], [x1, y1]);
                 if (!result) return console.log("unable to perform attack");
                 setLockedTx(result.tx);
-                eventBus.dispatchEvent({ type: "sui", action: "aim:success" });
+                eventBus.dispatchEvent({ type: "sui", action: "aim", message: "Unit can shoot" });
             }
         }
 
@@ -82,9 +99,14 @@ export function Playground() {
                 if (!res) return console.log("unable to end turn");
 
                 await executeTransaction(res.tx)!.wait();
-                eventBus.dispatchEvent({ type: "sui", action: "next_turn:success" });
                 if (map) {
-                    map.map.map.turn += 1;
+                    const turn = ++map.map.map.turn;
+                    eventBus.dispatchEvent({
+                        type: "sui",
+                        action: "next_turn",
+                        message: "Next turn: " + turn,
+                        turn,
+                    });
                 }
             }
         }
@@ -98,21 +120,59 @@ export function Playground() {
         };
     }, [map, executeTransaction]);
 
-    useEffect(() => {
-        if (!lockedTx) return;
+    /**
+     * Submit the locked transaction when the user clicks the confirm button.
+     * This event listener is reset every time a transaction is executed.
+     */
+    useEffect(
+        function submitLockedTransaction() {
+            if (!lockedTx) return;
 
-        /** Listen to the ui:confirm event, trigger tx sending */
-        async function onConfirm(event: GameEvent["ui"]) {
-            if (event.action === "confirm" && lockedTx && !isExecuting) {
-                setLockedTx(null);
-                const res = await executeTransaction(lockedTx)!.wait();
-                eventBus.dispatchEvent({ type: "sui", action: "tx:success", effects: res.effects });
+            /** Listen to the ui:confirm event, trigger tx sending */
+            async function onConfirm(event: GameEvent["ui"]) {
+                if (event.action === "confirm" && lockedTx && canTransact) {
+                    setLockedTx(null);
+                    const res = await executeTransaction(lockedTx)!.wait();
+                    const events = res.events || [];
+                    eventBus.dispatchEvent({
+                        type: "sui",
+                        action: "tx:success",
+                        effects: res.effects,
+                    });
+
+                    events.forEach((event) => {
+                        switch (event.type) {
+                            case `${packageId}::event::AttackEvent`:
+                                const data = event.parsedJson as AttackEvent;
+                                eventBus.dispatchEvent({
+                                    type: "sui",
+                                    action: "attack",
+                                    message: printAttackEvent(data),
+                                    data,
+                                });
+                                break;
+                            case `${packageId}::event::MoveEvent`: {
+                                const { path } = event.parsedJson as { path: [number, number][] };
+                                eventBus.dispatchEvent({
+                                    type: "sui",
+                                    action: "path",
+                                    path,
+                                    message:
+                                        "Unit moves: " +
+                                        path.map((p) => `(${p.toString()})`).join(", "),
+                                });
+                                break;
+                            }
+                        }
+                    });
+                }
             }
-        }
 
-        eventBus.addEventListener("ui", onConfirm);
-        return () => eventBus.removeEventListener("ui", onConfirm);
-    }, [lockedTx]);
+            eventBus.addEventListener("ui", onConfirm);
+            return () => eventBus.removeEventListener("ui", onConfirm);
+        },
+        [lockedTx, map],
+    );
 
     const centerDiv = (children: any) => (
         <div className="text-center bg-black/40 w-full h-full flex items-center justify-center">
@@ -125,9 +185,12 @@ export function Playground() {
     if (!modelsLoaded) return centerDiv("Models not loaded");
     if (!map)
         return centerDiv(
-            <div className="block">
+            <div className="block text-md">
                 <h1 className="mb-10">Map not found</h1>
                 <button onClick={() => createMap()}>create demo 1</button>
+                <NavLink className="menu-control mt-10" to="/commander">
+                    Back
+                </NavLink>
             </div>,
         );
 
@@ -146,10 +209,10 @@ export function Playground() {
         </>
     );
 
+    // === Transaction functions ===
+
     async function createMap() {
-        if (isExecuting) return;
-        if (!zkLogin.address) return;
-        if (!executeTransaction) return;
+        if (!canTransact) return;
 
         const tx = new Transaction();
         const r1 = tx.moveCall({ target: `${packageId}::recruit::default` });
@@ -186,9 +249,7 @@ export function Playground() {
     /** Perform ranged attack. I've been waiting for this soooo long */
     async function performAttack(unit: [number, number], target: [number, number]) {
         if (!map) return;
-        if (isExecuting) return;
-        if (!zkLogin.address) return;
-        if (!executeTransaction) return;
+        if (!canTransact) return;
 
         const tx = new Transaction();
         const game = tx.sharedObjectRef({
@@ -211,7 +272,7 @@ export function Playground() {
             ],
         });
 
-        tx.setSender(zkLogin.address);
+        tx.setSender(zkLogin.address!);
 
         const res = await client.dryRunTransactionBlock({
             transactionBlock: await tx.build({ client: client as any }),
@@ -223,9 +284,7 @@ export function Playground() {
     /** Move the unit on the map. The initial coordinate is the gridPosition of the unit. */
     async function moveUnit(path: [number, number][]) {
         if (!map) return;
-        if (isExecuting) return;
-        if (!zkLogin.address) return;
-        if (!executeTransaction) return;
+        if (!canTransact) return;
 
         const tx = new Transaction();
         const game = tx.sharedObjectRef({
@@ -237,7 +296,7 @@ export function Playground() {
         const pathArg = tx.pure(bcs.vector(bcs.vector(bcs.u16())).serialize(path));
 
         tx.moveCall({ target: `${packageId}::commander::move_unit`, arguments: [game, pathArg] });
-        tx.setSender(zkLogin.address);
+        tx.setSender(zkLogin.address!);
 
         const res = await client.dryRunTransactionBlock({
             transactionBlock: await tx.build({ client: client as any }),
@@ -249,9 +308,7 @@ export function Playground() {
     /** End the turn. */
     async function nextTurn() {
         if (!map) return;
-        if (isExecuting) return;
-        if (!zkLogin.address) return;
-        if (!executeTransaction) return;
+        if (!canTransact) return;
 
         const tx = new Transaction();
         const game = tx.sharedObjectRef({
@@ -261,7 +318,7 @@ export function Playground() {
         });
 
         tx.moveCall({ target: `${packageId}::commander::next_turn`, arguments: [game] });
-        tx.setSender(zkLogin.address);
+        tx.setSender(zkLogin.address!);
 
         const res = await client.dryRunTransactionBlock({
             transactionBlock: await tx.build({ client: client as any }),
@@ -314,8 +371,14 @@ export function GameApp({
             }
         });
 
-        eventBus.addEventListener("sui", ({ action }) => {
-            if (action == "next_turn:success") {
+        eventBus.addEventListener("sui", ({ action, data }) => {
+            if (action == "attack") {
+                let params = data as AttackEvent;
+                game.applyAttackEvent(params);
+                console.log("game should apply attack", params);
+            }
+
+            if (action == "next_turn") {
                 game.turn = game.turn + 1;
             }
         });
@@ -359,7 +422,7 @@ const LOG_LENGTH = 5;
 export function UI({
     eventBus,
     isExecuting,
-    turn,
+    turn: initilTurn,
 }: {
     eventBus: EventBus;
     isExecuting?: boolean;
@@ -377,9 +440,16 @@ export function UI({
 
     const [panelDisabled, setPanelDisabled] = useState(true);
     const [shootMode, setShootMode] = useState(false);
+    const [turn, setTurn] = useState(initilTurn);
     const [log, setLog] = useState<string[]>([]);
 
     useEffect(() => {
+        eventBus.addEventListener("sui", (event) => {
+            if (event.action === "next_turn") {
+                setTurn(event.turn);
+            }
+        });
+
         eventBus.addEventListener("game", (event) => {
             if (event.action === "mode_switch") {
                 if (event.mode instanceof ShootMode) setShootMode(true);
@@ -392,8 +462,8 @@ export function UI({
 
         eventBus.all((event) => {
             setLog((log) => {
-                let { type, action } = event;
-                let fullLog = [...log, `${type}: ${action ? action : ""}`];
+                let { type, action, message } = event;
+                let fullLog = [...log, `${type}: ${message || action || ""}`];
                 if (fullLog.length > LOG_LENGTH) fullLog = fullLog.slice(fullLog.length - 5);
                 return fullLog;
             });
@@ -435,6 +505,15 @@ export function UI({
                 {button("confirm", panelDisabled)}
                 {button("cancel", panelDisabled)}
                 {button("edit")}
+                <button
+                    onClick={() => {
+                        localStorage.removeItem(LS_KEY);
+                        window.location.reload();
+                    }}
+                    className={"action-button mt-40" + (isExecuting ? " disabled" : "")}
+                >
+                    Exit
+                </button>
             </div>
         </div>
     );
@@ -449,4 +528,15 @@ function loadCamera() {
     const camera = new Camera(75, aspect, 0.1, 100);
     camera.resetForSize(SIZE);
     return camera;
+}
+
+/**
+ * Print the attack event as a nice string.
+ */
+function printAttackEvent(event: AttackEvent): string {
+    if (event.is_missed) return "Miss!";
+    if (event.is_dodged) return "Dodged!";
+    if (event.is_kia) return "Killed in action! Damage: " + event.damage;
+    if (event.is_crit) return `Critical hit! ${event.damage} damage`;
+    return `Dealt ${event.damage} damage`;
 }
