+++ app/src/constants.ts
@@ -11,4 +11,4 @@ export const CHARACTER_BUILDER_ID =
 export const COMMANDER_PACKAGE_ID =
     "0x6a7c6106c49e77cbc376763613855ae66999d04f243acc8b9c34293436c50988";
 export const COMMANDER_V2_PACKAGE_ID =
-    "0xb55eccd9d0ee185411d9bbccb8830ae6d8aaf7ce30ce43c69689f9824427a2ec";
+    "0xfad4a80c5ef4adef1820dadadee5349caf277f9eb4154a3fc94621d443d02fd7";
