+++ packages/commander-v2/tests/commander_tests.move
@@ -2,5 +2,3 @@
 // SPDX-License-Identifier: MIT
 
 module commander::commander_tests;
-
-// ...
